% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_country_name.R
\name{add_country_name}
\alias{add_country_name}
\title{Convert country ISO3c codes to name column}
\usage{
add_country_name(data, col, country.col.name = NA)
}
\arguments{
\item{data}{dataframe. Must contain column of ISO3c codes}

\item{col}{character. Column name within data that cont}

\item{country.col.name}{character. User specified name of new country name column. Default will take input col and add "_name" for the new column name.}
}
\description{
`add_country_name()` takes an ISO 3166-1 alpha-3 codes column and creates a new column with full country names using OWID table.
}
\examples{

# Convert importer column codes to new column of full names
glimpse(mini_artis \%>\% 
          add_country_name("importer_iso3c"))

# Provide new column name

glimpse(mini_artis \%>\% 
          add_country_name("importer_iso3c",
                            country.col.name = "test_column_name"))

}
