% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_region.R
\name{add_region}
\alias{add_region}
\title{Convert country ISO3c codes to region name column}
\usage{
add_region(data, col, region.col.name = NA)
}
\arguments{
\item{data}{dataframe. Must contain chr column of ISO3c codes}

\item{col}{character. Column name within data that contains ISO3c codes.}

\item{region.col.name}{character. User specified name of new region name column. Default will take input col and add "_name" for the new column name.}
}
\description{
`add_region()` takes an ISO 3166-1 alpha-3 codes column and creates a new column with corresponding region names using OWID table.
}
\examples{

# Convert importer column codes to new column of region names
glimpse(mini_artis \%>\% 
          add_region("importer_iso3c"))
          
# Provide new column name
glimpse(mini_artis \%>\% 
          add_region("importer_iso3c",
                      region.col.name = "test_column_name"))

}
