% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chord.R, R/plot_chord_2.R
\name{plot_chord}
\alias{plot_chord}
\title{Function that creates a chord diagram from ARTIS data.}
\usage{
plot_chord(
  data,
  focal_country = NA,
  species = NA,
  years = NA,
  producers = NA,
  exporters = NA,
  importers = NA,
  hs_codes = NA,
  prod_method = NA,
  prod_environment = NA,
  export_source = NA,
  weight = "live",
  plot_region = FALSE,
  region_colors = region6_palette
)

plot_chord(
  data,
  focal_country = NA,
  species = NA,
  years = NA,
  producers = NA,
  exporters = NA,
  importers = NA,
  hs_codes = NA,
  prod_method = NA,
  prod_environment = NA,
  export_source = NA,
  weight = "live",
  plot_region = FALSE,
  region_colors = region6_palette
)
}
\arguments{
\item{species}{list of species/species groups to include, default NA - includes all species.}

\item{years}{list of years to include, default NA - includes all years.}

\item{producers}{list of producers (as iso3 codes) to include, default NA - includes all producers.}

\item{exporters}{list of exporters (as iso3 codes) to include, default NA - includes all exporters.}

\item{importers}{list of importers (as iso3 codes) to include, default NA - includes all importers.}

\item{hs_codes}{list of hs level 6 codes to include, default NA - includes all hs6 codes.}

\item{prod_method}{list of production methods (capture, aquaculture, or unknown), default NA - includes all production methods.}

\item{prod_environment}{list of environments (marine, inland, or unknown), default NA - includes all environments}

\item{export_source}{list of types of export (domestic export, foreign export, or error export), default NA - all export sources.}

\item{weight}{trade quantity type to visualize ("live" for live weight or "product" for product weight), default "live."}
}
\description{
Function that creates a chord diagram from ARTIS data.

Function that creates a chord diagram from ARTIS data.
}
