% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_owid_regions.R
\name{convert_owid_regions}
\alias{convert_owid_regions}
\title{Transform ARTIS from country to regional data}
\usage{
convert_owid_regions(df)
}
\arguments{
\item{df}{- An ARTIS dataframe (needs to include source_country_iso3c, exporter_iso3c, importer_iso3c)}
}
\description{
Function that takes an ARTIS dataframe and summarizes source, export and import countries as regions.
}
