% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{ARTIS Bar Chart}
\usage{
plot_bar(
  data,
  bar_group,
  value = "live_weight_t",
  fill_type = NA,
  top_n = 10,
  y.lab = "",
  x.lab = NA,
  fill.lab = "",
  plot.title = "",
  facet_variable = NA,
  facet_n = NA
)
}
\arguments{
\item{data}{dataframe. ARTIS or consumption data.}

\item{bar_group}{character. Column name in data to group data into bars in plot (e.g. "sciname").}

\item{value}{character. Column name in data for the values the bar length.}

\item{fill_type}{character. Column name to use to stack the bars (e.g. "method" or "habitat").}

\item{top_n}{numeric. Number of top ranked groups to plot.}

\item{y.lab}{character. Text for the y-axis label (e.g. "My title").}

\item{x.lab}{character. Text for the x-axis label.}

\item{fill.lab}{character. Text for the fill legend label.}

\item{plot.title}{character. Text for the plot title.}

\item{facet_variable}{character. Column name to facet by.}

\item{facet_n}{numeric. Number of facets to include. Must be specified if a facet_variable is specified.}
}
\description{
Function that creates a horizontal bar chart for ARTIS or consumption data.
}
\examples{
# Use `mini_artis` dataframe included in package

# Basic plot with importer columns
plot_bar(mini_artis, 
        bar_group = "importer_iso3c")

# Importer bars filled by "habitat" grouping 
plot_bar(mini_artis, 
        bar_group = "importer_iso3c", 
        fill_type = "habitat")

# Importer bars faceted by "habitat"
plot_bar(mini_artis, 
        bar_group = "importer_iso3c", 
        facet_variable  = "habitat", 
        facet_n = 3)

}
