% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Creates a chloropleth map with trade flows of ARTIS data}
\usage{
plot_map(
  data,
  value = "live_weight_t",
  country_fill = NA,
  flow_arrows = FALSE,
  n_flows = 10,
  arrow_label = NA,
  fill_label = NA,
  caption_label = NA,
  regions = FALSE
)
}
\arguments{
\item{data}{an ARTIS dataframe.}

\item{value}{the numeric variable for the fill/color. Default is live_weight_t.}

\item{country_fill}{optional variable by which to color countries - either "import" to color by total imports or export to color by total export.}

\item{flow_arrows}{logical variable - set to TRUE to include top trade flow arrows.}

\item{n_flows}{number of top trade flow arrows to include if flow_arrows = TRUE - defaults to 10.}
}
\value{
None
}
\description{
This is a function that creates a chloropleth map (optional) with trade flow arrows (optional)  based on the ARTIS dataset.
}
