% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partner_line.R
\name{plot_partner_line}
\alias{plot_partner_line}
\title{Creates a line graph of ARTIS data by trade partner}
\usage{
plot_partner_line(
  data,
  trade_flow,
  prop_flow_cutoff = 0.05,
  species = NA,
  years = NA,
  producers = NA,
  exporters = NA,
  importers = NA,
  hs_codes = NA,
  prod_method = NA,
  prod_environment = NA,
  export_source = NA,
  weight = "live",
  plot.title = ""
)
}
\arguments{
\item{data}{an ARTIS dataframe.}

\item{trade_flow}{trade partner type interested in stacking. Default = "export" - plots export partners to the focal region. "import" - plots import partners from focal region.}

\item{prop_flow_cutoff}{default prop_flow_cutoff = 0.05 means trade volumes that comprise less than 5\% of the total trade are lumped together as "Other".}

\item{species}{list of species/species groups to include, default NA - includes all species.}

\item{years}{list of years to include, default NA - includes all years.}

\item{producers}{list of producers to include, default NA - includes all producers.}

\item{exporters}{list of exporters to include, default NA - includes all exporters.}

\item{importers}{list of importers to include, default NA - includes all importers.}

\item{hs_codes}{list of hs level 6 codes to include, default NA - includes all hs6 codes.}

\item{prod_method}{list of production methods (capture, aquaculture, etc), default NA - includes all production methods.}

\item{prod_environment}{list of environments (marine, inland, etc), default NA - includes all environments}

\item{export_source}{list of types of export (domestic export, foreign export, error export), default NA - all export sources.}

\item{weight}{trade quantity type to visualize, default "live" for live weight, otherwise product weight}

\item{plot.title}{title for plot generated}
}
\value{
Line graph of ARTIS data by trade partner.
}
\description{
This is a function that creates a line graph showcasing seafood trade volumes by trading partner.
}
