% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Creates a Sankey plot of ARTIS data}
\usage{
plot_sankey(
  data,
  cols = c("source_country_iso3c", "exporter_iso3c", "importer_iso3c"),
  prop_flow_cutoff = 0.05,
  weight = "live_weight_t",
  show.other = FALSE,
  plot.title = ""
)
}
\arguments{
\item{data}{an ARTIS dataframe.}

\item{cols}{ADD DESCRIPTION}

\item{prop_flow_cutoff}{default prop_flow_cutoff = 0.05 means trade volumes that comprise less than 5\% of the total trade are lumped together as "Other".}

\item{weight}{trade quantity type to visualize ("live" for live weight or "product" for product weight), default "live."}
}
\value{
None
}
\description{
This is a function that creates a sankey plot showcasing seafood supply chains from producer to importer in the ARTIS dataset.
}
