% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Creates a Sankey plot of ARTIS data}
\usage{
plot_sankey(
  data,
  cols = c("source_country_iso3c", "exporter_iso3c", "importer_iso3c"),
  prop_flow_cutoff = 0.05,
  value = "live_weight_t",
  show.other = TRUE,
  plot.title = ""
)
}
\arguments{
\item{data}{dataframe. An ARTIS trade or consumption dataframe.}

\item{cols}{vector. Column names to generate the sections of the Sankey plot, in the order they should appear (left to right).}

\item{prop_flow_cutoff}{integer. A percent in which trade volumes that comprise less than x\% of the total trade are renamed as "Other". Default prop_flow_cutoff = 0.05 means trade volumes less than 5% are labeled as "Other".}

\item{value}{character. Trade quantity column name to visualize. Default is "live_weight_t"}

\item{show.other}{logical. Controls whether or not nodes within a column falling below the prop_flow_cutoff threshold should be displayed in a group ("Other"). Default value is TRUE, filtering for threshold occurs regardless is "Other" is displayed.}

\item{plot.title}{character. User-specified title to the plot. Default is a blank title.}
}
\value{
None
}
\description{
This is a function that creates a sankey plot showcasing elements of seafood supply chains in ARTIS or consumption datasets.
}
\examples{
# Use `mini_artis` dataframe included in package

# basic default sankey
plot_sankey(mini_artis)

# define alternative columns to use in sankey plot
plot_sankey(mini_artis, 
           cols = c("method", "habitat"))

# Do not filter value by proportion cutoff - retain all flows
plot_sankey(mini_artis,
           prop_flow_cutoff = 0)

}
