% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Creates a time series line graph/stacked area graph}
\usage{
plot_ts(
  data,
  artis_var = NA,
  trade_flow = NA,
  prop_flow_cutoff = 0.05,
  species = NA,
  years = NA,
  producers = NA,
  exporters = NA,
  importers = NA,
  regions = NA,
  hs_codes = NA,
  prod_method = NA,
  prod_environment = NA,
  export_source = NA,
  region_self_loops = TRUE,
  weight = "live_weight_t",
  facet_variable = NA,
  facet_values = NA,
  quantity.lab = NA,
  plot.type = "line",
  legend.title = NA,
  plot.title = ""
)
}
\description{
Creates a time series line graph/stacked area graph
}
